requirements.txt
aiohttp==3.12.15
discord.py==2.6.3
protobuf==6.32.1
python-dotenv==1.1.1
PyYAML==6.0.3
